Refresh token (#12)
* Add "expires_at" to token data if applicable

If the expires_in key exists the expires_at will be added. This can be
useful to determine if the token needs refreshing without a failing call
to the api.

The expires_at is an absolute time when the token expires.

fixup

* Add support for refreshing oauth token

The "oauth.refresh" message is now handled and will trigger a token
refresh if possible. The Message data must contain valid "skill_id" and
"app_id" fields.

* gitignore: Add vim cache files